requirements.txt
PyOpenGL>=3.1.6
pygame>=2.5.0
